Split code into required files
