+++ src/main/java/mael/Parser.java
@@ -1,0 +1,5 @@
+package mael;
+
+class Parser {
+    
+}
